Remove invalid username / password work around from multibranch pipeline job
Previously there was a bug in the pipeline support which required a
credential definition for a job, even if the repository did not require
credentials.  That bug is fixed, and this confirms it.

This change also updates the job to use the latest scm-api, and replaces a reference to
icon com.github.mjdetullio.jenkins.plugins.multibranch.BallColorFolderIcon from multi-branch-project-plugin@0.5.1
with a reference to
icon jenkins.branch.MetadataActionFolderIcon from branch-api@2.0.2.
That may have caused a loss of functionality in another area. That needs
more investigation.